+++ packages/moonstone/ExpandableList/ExpandableList.js
@@ -379,6 +379,12 @@ const ExpandableList = Pure(
 		children: compareChildren
 	}},
 	Expandable(
+		{getSelectedNode: (node, {selected}) => {
+			const _selected = selected ? selected : 0;
+			const selectedIndex = Array.isArray(_selected) && _selected.length ? _selected[0] : _selected;
+			const selectedNode = node ? node.querySelector('[data-index="' + selectedIndex + '"]') : null;
+			return selectedNode;
+		}},
 		Changeable(
 			{change: 'onSelect', prop: 'selected'},
 			ExpandableListBase